Add gen-movie-catalog.py
Add gen-movie-catalog.py to generate a catalog of movie information by
scanning the local file system for movies. It tries to predict the movie
name from the movie file name and then fetches imdb information for that
movie using omdb api (www.omdbapi.com). It can either generate csv or
json output. It can also accept previously generated json output and
convert it to required CSV.

Currently all configuration and program modes are embedded in script and
the script needs to be edited in order to switch functionality. Later we
can make it accept outside config and command line parameters if need
be.

This is just the initial version of code and has lots of TODO markers in
the code.