Add plot-file-size python script
Plots a PDF of all the file sizes within a directory by recursively
walking the directory. This code demostrates the use of simple
probability functions and plotting in python.

For beter plotting directories containing movies or songs, it ignores
some types of files with extensions specified in the invalidext set.